award badge state icons

+++ app/assets/javascripts/angular/directives/badges/awardEarnedBadges.js.coffee
@@ -4,15 +4,17 @@
     vm.BadgeService = BadgeService
     BadgeService.getBadges(vm.studentId)
 
-    vm.badgeActionable = (badge)->
-      vm.badgeEarnedForGrade(badge) ||
-      badge.can_earn_multiple_times ||
-      badge.earned_badge_count < 1
-
     vm.badgeEarnedForGrade = (badge)->
       BadgeService.studentEarnedBadgeForGrade(vm.studentId, badge.id, vm.gradeId)
 
+    vm.badgeAvailable = (badge)->
+      badge.can_earn_multiple_times || badge.earned_badge_count < 1
+
+    vm.badgeAwardable = (badge)->
+      vm.badgeAvailable(badge) && !vm.badgeEarnedForGrade(badge)
+
     vm.awardBadge = (badge)->
+      return if !vm.badgeAvailable(badge)
       if earnedBadge = vm.badgeEarnedForGrade(badge)
         BadgeService.deleteEarnedBadge(earnedBadge)
       else
